+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [12.6.1] - 2024-04-30
+### Fixed
+- Maintenance release due to issue occurred during publish of 12.6.0
+
 ## [12.6.0] - 2024-04-30
 ### Added
 - **feat**: array/groupBy, (suggested by [Xander Berkein](https://github.com/xanderberkein) eg: